create custom hook that uses local storage manage todos
